rename tofworkflowutil dir
