+++ Detectors/GlobalTrackingWorkflow/tofworkflow/src/tof-matcher-tpc.cxx
@@ -14,7 +14,7 @@
 /// @brief  Basic DPL workflow for TOF reconstruction starting from digits
 
 #include "DetectorsBase/Propagator.h"
-#include "TOFWorkflow/ClusterReaderSpec.h"
+#include "TOFWorkflowUtils/ClusterReaderSpec.h"
 #include "TOFWorkflow/TOFMatchedWriterSpec.h"
 #include "TOFWorkflow/TOFCalibWriterSpec.h"
 #include "Framework/WorkflowSpec.h"
