+++ Detectors/GlobalTrackingWorkflow/tofworkflow/src/tof-reco-workflow.cxx
@@ -15,17 +15,17 @@
 
 #include "DetectorsBase/Propagator.h"
 #include "GlobalTrackingWorkflow/TrackTPCITSReaderSpec.h"
-#include "TOFWorkflow/DigitReaderSpec.h"
-#include "TOFWorkflow/TOFDigitWriterSpec.h"
-#include "TOFWorkflow/ClusterReaderSpec.h"
-#include "TOFWorkflow/TOFClusterizerSpec.h"
-#include "TOFWorkflow/TOFClusterWriterSpec.h"
+#include "TOFWorkflowUtils/DigitReaderSpec.h"
+#include "TOFWorkflowUtils/TOFDigitWriterSpec.h"
+#include "TOFWorkflowUtils/ClusterReaderSpec.h"
+#include "TOFWorkflowUtils/TOFClusterizerSpec.h"
+#include "TOFWorkflowUtils/TOFClusterWriterSpec.h"
 #include "TOFWorkflow/TOFMatchedWriterSpec.h"
 #include "TOFWorkflow/TOFCalibWriterSpec.h"
-#include "TOFWorkflow/TOFRawWriterSpec.h"
-#include "TOFWorkflow/CompressedDecodingTask.h"
-#include "TOFWorkflow/EntropyEncoderSpec.h"
-#include "TOFWorkflow/EntropyDecoderSpec.h"
+#include "TOFWorkflowUtils/TOFRawWriterSpec.h"
+#include "TOFWorkflowUtils/CompressedDecodingTask.h"
+#include "TOFWorkflowUtils/EntropyEncoderSpec.h"
+#include "TOFWorkflowUtils/EntropyDecoderSpec.h"
 #include "Framework/WorkflowSpec.h"
 #include "Framework/ConfigParamSpec.h"
 #include "TOFWorkflow/RecoWorkflowSpec.h"
