+++ Detectors/TOF/workflow/CMakeLists.txt
@@ -25,5 +25,5 @@ o2_add_library(TOFWorkflowUtils
 o2_add_executable(entropy-encoder-workflow
                   SOURCES src/entropy-encoder-workflow.cxx
                   COMPONENT_NAME tof
-                  PUBLIC_LINK_LIBRARIES O2::TOFWorkflow)
+                  PUBLIC_LINK_LIBRARIES O2::TOFWorkflowUtils)
