+++ Detectors/TOF/workflow/include/TOFWorkflowUtils/CompressedInspectorTask.h
