+++ Detectors/TOF/workflow/include/TOFWorkflowUtils/EntropyEncoderSpec.h
