+++ Detectors/TOF/workflow/src/DigitReaderSpec.cxx
@@ -17,7 +17,7 @@
 #include "Framework/ControlService.h"
 #include "Framework/ConfigParamRegistry.h"
 #include "Framework/Logger.h"
-#include "TOFWorkflow/DigitReaderSpec.h"
+#include "TOFWorkflowUtils/DigitReaderSpec.h"
 #include "DataFormatsParameters/GRPObject.h"
 
 using namespace o2::framework;
