+++ Detectors/TOF/workflow/src/EntropyEncoderSpec.cxx
@@ -15,7 +15,7 @@
 #include "Framework/ControlService.h"
 #include "Framework/ConfigParamRegistry.h"
 #include "TOFBase/Digit.h"
-#include "TOFWorkflow/EntropyEncoderSpec.h"
+#include "TOFWorkflowUtils/EntropyEncoderSpec.h"
 
 using namespace o2::framework;
