+++ Detectors/TOF/workflow/src/entropy-encoder-workflow.cxx
@@ -8,7 +8,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include "TOFWorkflow/EntropyEncoderSpec.h"
+#include "TOFWorkflowUtils/EntropyEncoderSpec.h"
 #include "CommonUtils/ConfigurableParam.h"
 #include "Framework/ConfigParamSpec.h"
